+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - If a resource fails while initializing, now all intermediate child resources are destroyed as well.
 - If a resource fails while being destroyed, now its child resources are destroyed as well.
+- Resources can no longer provision child resources after teardown. This closes a loophole where resources could escape destruction.
 
 ## [0.4.1] - 2023-01-14